refactor(activities) :hammer: refactor in activities
