+++ app/src/main/AndroidManifest.xml
@@ -25,7 +25,7 @@
         android:usesCleartextTraffic="true">
 
         <activity
-            android:name=".routes.activity.splashscreen.SplashActivity"
+            android:name=".routes.activity.splash_screen.SplashActivity"
             android:exported="true"
             android:label="@string/app_name"
             android:screenOrientation="portrait"
@@ -59,7 +59,7 @@
         </activity>
 
         <activity
-            android:name=".routes.activity.PictureDetailActivity"
+            android:name=".routes.activity.picture_detail.PictureDetailActivity"
             android:exported="true"
             android:label="@string/app_name"
             android:screenOrientation="portrait"
@@ -70,7 +70,7 @@
         </activity>
 
         <activity
-            android:name=".routes.activity.CameraActivity"
+            android:name=".routes.activity.cameraV1.CameraActivity"
             android:exported="true"
             android:label="@string/app_name"
             android:screenOrientation="portrait"
@@ -81,7 +81,7 @@
         </activity>
 
         <activity
-            android:name=".routes.activity.ConfigurationsActivity"
+            android:name=".routes.activity.configuration.ConfigurationsActivity"
             android:exported="true"
             android:label="@string/app_name"
             android:screenOrientation="portrait"
@@ -92,7 +92,7 @@
         </activity>
 
         <activity
-            android:name=".routes.activity.CameraProV2Activity"
+            android:name=".routes.activity.cameraV2.CameraProV2Activity"
             android:exported="true"
             android:label="@string/app_name"
             android:screenOrientation="portrait"
