+++ app/src/main/java/com/iiap/deeplarva/helpers/pictureInputHelper/PictureByCameraProHandler.kt
@@ -9,8 +9,8 @@ import android.widget.Toast
 import androidx.annotation.RequiresApi
 import com.iiap.deeplarva.domain.constants.AppConstants
 import com.iiap.deeplarva.domain.constants.SharedPreferencesConstants
-import com.iiap.deeplarva.routes.activity.CameraActivity
-import com.iiap.deeplarva.routes.activity.CameraProV2Activity
+import com.iiap.deeplarva.routes.activity.cameraV1.CameraActivity
+import com.iiap.deeplarva.routes.activity.cameraV2.CameraProV2Activity
 import com.iiap.deeplarva.utils.BitmapUtils
 import com.iiap.deeplarva.utils.PreferencesHelper
