+++ app/src/main/java/com/iiap/deeplarva/routes/activity/cameraV1/CameraActivityView.kt
@@ -1,4 +1,4 @@
-package com.iiap.deeplarva.routes.activity.view
+package com.iiap.deeplarva.routes.activity.cameraV1
 
 import androidx.appcompat.app.AppCompatActivity
 import androidx.constraintlayout.widget.ConstraintLayout
