+++ app/src/main/java/com/iiap/deeplarva/routes/activity/cameraV1/ICameraViewListener.kt
@@ -1,4 +1,4 @@
-package com.iiap.deeplarva.routes.activity.view
+package com.iiap.deeplarva.routes.activity.cameraV1
 
 interface ICameraViewListener {
     // Behavior of View
