+++ app/src/main/java/com/iiap/deeplarva/routes/activity/cameraV2/CameraProV2Activity.kt
@@ -1,4 +1,4 @@
-package com.iiap.deeplarva.routes.activity
+package com.iiap.deeplarva.routes.activity.cameraV2
 
 import android.content.Intent
 import android.os.Build
@@ -12,13 +12,12 @@ import androidx.lifecycle.ViewModelProvider
 import com.iiap.deeplarva.databinding.ActivityCameraPro2Binding
 import com.iiap.deeplarva.domain.constants.AppConstants
 import com.iiap.deeplarva.domain.constants.SharedPreferencesConstants
-import com.iiap.deeplarva.utils.PreferencesHelper
 import com.iiap.deeplarva.modules.camerapro2.infraestructure.SensitivityProvider
-import com.iiap.deeplarva.routes.activity.observables.CameraV2Model
-import com.iiap.deeplarva.routes.activity.stores.CameraParameterStore
+import com.iiap.deeplarva.modules.CameraParameterStore
 import com.iiap.deeplarva.ui.widget.dialogs.SeekDialog
 import com.iiap.deeplarva.ui.widget.dialogs.SelectableDialog
 import com.iiap.deeplarva.ui.widget.dialogs.ShooterSpeedDialog
+import com.iiap.deeplarva.utils.PreferencesHelper
 import com.iiap.deeplarva.utils.SpeedUtils
 import com.kylecorry.andromeda.camera.ImageCaptureSettings
 import com.kylecorry.andromeda.core.math.DecimalFormatter
