+++ app/src/main/java/com/iiap/deeplarva/routes/activity/configuration/ConfigurationsActivity.kt
@@ -1,4 +1,4 @@
-package com.iiap.deeplarva.routes.activity
+package com.iiap.deeplarva.routes.activity.configuration
 
 import android.os.Bundle
 import android.view.MenuItem
