+++ app/src/main/java/com/iiap/deeplarva/routes/activity/main/PictureActivityView.kt
@@ -13,8 +13,7 @@ import com.iiap.deeplarva.domain.constants.AppConstants
 import com.iiap.deeplarva.domain.constants.PermissionsConstans
 import com.iiap.deeplarva.domain.view.PictureListEntity
 import com.iiap.deeplarva.helpers.pictureInputHelper.PictureInputHelper
-import com.iiap.deeplarva.routes.activity.PictureDetailActivity
-import com.iiap.deeplarva.routes.activity.view.IPictureViewListener
+import com.iiap.deeplarva.routes.activity.picture_detail.PictureDetailActivity
 import com.iiap.deeplarva.ui.adapter.PictureAdapterList
 import com.iiap.deeplarva.ui.adapter.PictureItemListListener
 import com.iiap.deeplarva.ui.widget.listHandler.ListEventListener
