+++ app/src/main/res/layout/activity_camera.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:id="@+id/clMain"
     android:background="@color/black"
-    tools:context=".routes.activity.CameraActivity">
+    tools:context=".routes.activity.cameraV1.CameraActivity">
 
     <RelativeLayout
         android:layout_width="match_parent"
