+++ app/src/main/res/layout/activity_picture_detail.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/activityBg"
-    tools:context=".routes.activity.PictureDetailActivity">
+    tools:context=".routes.activity.picture_detail.PictureDetailActivity">
 
     <com.google.android.material.appbar.AppBarLayout
         android:id="@+id/appBarPictureDetail"
